loads last game to cookies and the inventory grabs the information from there

+++ server/models/playerMenu.js
@@ -1,72 +1,38 @@
-const jwt = require('jsonwebtoken')
-
-const config = require('../config/authentication')
+const { verifyTokenFromCookies } = require('../models/cookies')
 const Games = require('../models/game')
 
 async function createGame(request, response) {
-    let cookies = request.headers.cookie
-    let token = cookies.slice(12, cookies.length)
-    let userId = jwt.verify(token, config.secret).userId
+    let userId = verifyTokenFromCookies(request, 'accessToken', 'userId')
 
-    try {
-        const game = {
-            userId: userId,
-            inventory: {
-                gaiaGun: false,
-                cartridge: [],
-                spaceshipPieces: [],
-                hardDrivePieces: []
-            },
-            puzzles: []
-        }
-
-        const newGame = new Games(game)
-        await newGame.save()
-
-        response.status(200).send('Game created successfully')
+    const game = {
+        userId: userId,
+        inventory: {
+            gaiaGun: false,
+            cartridge: [],
+            spaceshipPieces: [],
+            hardDrivePieces: []
+        },
+        puzzles: []
     }
-    catch (error) {
-        console.log(error)
-    }
-}
 
-async function loadGame(request, response, next) {
-    let cookies = request.headers.cookie
-    let token = cookies.slice(12, cookies.length)
-    let userId = jwt.verify(token, config.secret).userId
+    const newGame = new Games(game)
+    await newGame.save()
 
-    await Games.find({ userId: userId })
-        .sort({ _id: -1 })
-        .limit(1)
-        .then((game) => {
-            try {
-                if (game.length === 0) {
-                    return response.status(404).send('You need to start a new game first')
-                }
-                next()
-            }
-            catch (error) {
-                console.log(error)
-            }
-        })
+    response.status(200).send('Game created')
 }
 
 async function deleteGame(request, response) {
-    let cookies = request.headers.cookie
-    let token = cookies.slice(12, cookies.length)
-    let userId = jwt.verify(token, config.secret).userId
+    let userId = verifyTokenFromCookies(request, 'accessToken', 'userId')
 
-    try {
-        await Games.deleteMany({ userId: userId })
-        response.status(200).send('Deleted games successfully')
+    if (!userId) {
+        return response.status(404).send('Game not found')
     }
-    catch (error) {
-        console.log(error)
-    }
+
+    await Games.deleteMany({ userId: userId })
+    response.status(200).send('Game deleted')
 }
 
 module.exports = {
     createGame,
-    loadGame,
     deleteGame,
 }
